+++ code/app/src/main/java/com/hamidat/nullpointersapp/moodClasses/moodHistory.java
@@ -1,4 +1,4 @@
-package com.hamidat.nullpointersapp;
+package com.hamidat.nullpointersapp.moodClasses;
 
 import android.util.Log;
